pyproject.toml
[project]
name = "contact"
version = "1.4.0"
description = "This Python curses client for Meshtastic is a terminal-based client designed to manage device settings, enable mesh chat communication, and handle configuration backups and restores."
authors = [
    {name = "Ben Lipsey",email = "ben@pdxlocations.com"}
]
license = "GPL-3.0-only"
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "meshtastic (>=2.6.0,<3.0.0)",
    "libsixel-python",
    "py-staticmaps @ git+https://github.com/flopp/py-staticmaps"
]

[project.urls]
Homepage = "https://github.com/pdxlocations/contact"
Issues = "https://github.com/pdxlocations/contact/issues"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
contact = "contact.__main__:start"
